app/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
numpy>=1.19.0
aiohttp>=3.8.0
moviepy>=1.0.3
oss2>=2.15.0
aliyun-python-sdk-core>=2.13.3
aliyun-python-sdk-nls-cloud-meta>=0.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
# YouTube下载和转录功能依赖
yt-dlp>=2023.12.30
websockets>=11.0.0
pytest>=7.0.0